+++ changelog.md
@@ -4,6 +4,7 @@
 - 16:00 - Admin section for Bookmarks tweaked.
 - 16:10 - `Bookmark` model extended with `is_read` field.
 - 16:50 - `Tag`s added to `Bookmark`s. Serializers adjusted.
+- 19:30 - Created basic Nuxt frontend, added Node to Compose.
 
 ### 06.04.2023, Thu